requirements.txt
django>=5.1,<6.0
requests>=2.32.0,<3.0
psycopg2-binary>=2.9,<3.0
django-allauth>=0.57.0,<1.0
PyJWT>=2.12.0,<3.0
cryptography>=46.0.0,<47.0
python-decouple>=3.8,<4.0
djangorestframework>=3.14,<4.0
dj-rest-auth>=5.0,<6.0
cloudinary>=1.40.0,<2.0
django-cloudinary-storage>=0.3.0,<1.0
Pillow>=10.0.0,<12.0
gunicorn>=22.0.0,<23.0
whitenoise>=6.8.0,<7.0

